backend/BLEO-backend/django-mongodb-api/requirements.txt
# pip install -r .\requirements.txt
django
djangorestframework
environs
pymongo